refactor: make incidencias section reusable
